Preemptively crashing if ExecutionThread-subclass leaves thread running
The method ExecutionThread::execute() is pure abstract, and subclasses
that attempts to call it after their destructor has finished will fail
horribly. This case is now detected and asserted upon from the
ExecutionThread, which will terminate the process via an assert if the
background thread is still running when its' destructor is called.