Exclusively handling lvalues in AST conversion
I ran into multiple issues, both with ANTLR and the CISH implementation
itself when dealing with lvalues in the grammar. All expressions are now
equal from the grammar's perspective, and lvalues are interpreted as
such when the AST is being generated. The implementation is now much
cleaner, but the compiler errors that will be generated for the users of
CISH are absolute bullshit and must be improved upon at some point. With
the current implementation, we have absolutely no information apart from
"I expected an Lvalue, but you gave me an Rvalue" - literally. I was
unable to bind statements/expressions to line numbers, remember? :)